Fixed all bugs in rectangle detection, still need to fix area and implement perimeter

+++ PolygonChecker/rectangleSolver.h
@@ -8,7 +8,7 @@ typedef struct pairs {
 
 bool isValidRectangle(PAIRS a, PAIRS b, PAIRS c, PAIRS d);  // e.l 
 
-int getLengthBetweenPoints(PAIRS, PAIRS);
+float getLengthBetweenPoints(PAIRS, PAIRS);
 
 PAIRS CreatePair(int, int); // M.B
